+++ src/tools/list.h
@@ -60,7 +60,7 @@ typedef struct
 
 list *create_list();
 int add_list_element(list *l, void *value);
-int get_list_size(list *l);
+size_t get_list_size(list *l);
 void *get_list_element(list *l, int list_index);
 bool list_contains_element(list *haystack, void *needle);
 void delete_list(list *l);
